+++ packages/javascript/lib/node/file/getFileExtension.ts
@@ -1,17 +1,8 @@
 import { extname } from 'path';
+import { getFileName } from './getFileName.js';
 
-export function getFileExtension(filePath: string, isWindows?: boolean): string | null {
-	let strPath: string | undefined = filePath.split(isWindows ? '\\' : '/').pop();
+export function getFileExtension(filePath: string): string | null {
+	const ext = extname(getFileName(filePath, true)).replace('.', '').toLowerCase();
 
-	if (!strPath) {
-		return null;
-	}
-
-	strPath = extname(strPath) || strPath;
-
-	if (strPath.indexOf('.') === -1) {
-		return null;
-	}
-
-	return strPath.split('.')?.pop()?.toLowerCase() || null;
+	return ext?.length > 0 ? ext : null;
 }