+++ packages/javascript/lib/node/file/getFileName.ts
@@ -1,22 +1,10 @@
 import { basename, extname, win32 } from 'path';
 
 export function getFileName(filePath: string, withExtension = false): string {
-	if (!filePath) {
-		return '';
-	}
-
+	// For Windows path
 	if (filePath.indexOf('/') === -1 && filePath.indexOf('\\') !== -1) {
-		// Windows path
-		if (withExtension) {
-			return win32.basename(filePath);
-		}
-
-		return win32.basename(filePath, extname(filePath));
+		return win32.basename(filePath, withExtension ? undefined : extname(filePath)) || '';
 	}
 
-	if (withExtension) {
-		return basename(filePath);
-	}
-
-	return basename(filePath, extname(filePath));
+	return basename(filePath, withExtension ? undefined : extname(filePath)) || '';
 }